Don't create a user if one with user ID exists
Programs that do things like getting the home dir end up getting the home dir of the other user by accident.

Use an existing user if the user ID and group ID match. Error out if the group ID matches, or if there are multiple
users with the same user ID.